Remove useless async, fix some code and loongarch64 pci virtio (#8)
* fix: remove useless async

* feat: update downcast-rs and fix some clippy warnnings

* fix: fix some diagnostics and useless async

* feat: remove useless deps and simplify some code

* feat: using PCI virtio with loongarch64

+++ filesystem/vfscore/src/lib.rs
@@ -63,7 +63,7 @@ pub type VfsResult<T> = core::result::Result<T, Errno>;
 pub trait BlockDevice: Send + Sync {
     fn read_block(&self, block: usize, buffer: &mut [u8]) -> VfsResult<usize>;
     fn write_block(&self, block: usize, buffer: &[u8]) -> VfsResult<usize>;
-    fn capacity(&self) -> VfsResult<usize>;
+    fn capacity(&self) -> VfsResult<u64>;
 }
 
 pub trait INodeInterface: DowncastSync + Send + Sync {
